LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/geolocation-API/enabled-by-permission-policy-attribute.https.sub-expected.txt

CONSOLE MESSAGE: Blocked https://127.0.0.1:9443/resources/testdriver.js from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Blocked https://127.0.0.1:9443/resources/testdriver-vendor.js from asking for credentials because it is a cross-origin request.

PASS Permissions policy "geolocation" can be enabled in same-origin iframe using allow="geolocation" attribute
PASS Permissions policy "geolocation" can be enabled in cross-origin iframe using allow="geolocation" attribute
